Update LICENSE to be public domain.
The readme says the software is in the public domain.

The LICENSE file says it is the MIT license.

**Neither** of these were added by the author of the repository.  However, The readme was recently updated by the author of the repository:

https://github.com/navdeep-G/setup.py/commit/46f7cd41a5d58adde653aae19c54cf689a3b2a91

From the above, I infer that the correct license should be the Unlicense license. I copied and pasted the text from the github template.  Presumably, it is the same as the below link.

https://unlicense.org/